refactor: refatorar Models de buyer, cnpj, order, provider e user

+++ back-end/src/migrations/20240731202222-buyer-migration.js
@@ -77,7 +77,7 @@ module.exports = {
         allowNull: false,
       },
       cnpjId: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.INTEGER(11),
       },
       confirm: {
         type: Sequelize.TINYINT(1),
